broke out udp functions
