+++ lib/tunneler_udp.h
@@ -1,0 +1,11 @@
+#ifndef ZITI_TUNNELER_SDK_TUNNELER_UDP_H
+#define ZITI_TUNNELER_SDK_TUNNELER_UDP_H
+
+#include "nf/ziti_tunneler.h"
+#include "lwip/ip_addr.h"
+#include "lwip/raw.h"
+
+extern u8_t recv_udp(void *tnlr_ctx_arg, struct raw_pcb *pcb, struct pbuf *p, const ip_addr_t *addr);
+extern void on_udp_packet(void *arg, struct udp_pcb *pcb, struct pbuf *p, const ip_addr_t *addr, u16_t port);
+
+#endif //ZITI_TUNNELER_SDK_TUNNELER_UDP_H
